네트워크 확인 움직임 확인
할일
Update
Draw

+++ Vector2D.h
@@ -6,7 +6,7 @@ public:
 	float x, y;
 
 	Vector2D() {}
-	Vector2D(float pX, float pY) :x((short)pX), y((short)pY) {}
+	Vector2D(float pX, float pY) :x(pX), y(pY) {}
 	Vector2D(short pX, short pY) :x(pX), y(pY) {}
 
 	Vector2D Normalize();
